nas-explorer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-magic==0.4.27
pdfplumber==0.11.0
python-docx==1.1.0
openpyxl==3.1.2
mutagen==1.47.0
Pillow==10.4.0
pdf2image==1.17.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
passlib[bcrypt]==1.7.4
smbprotocol==1.14.0
cryptography>=42.0.0
